add mocha MAP for cityes
