+++ src/components/app/app.jsx
@@ -22,7 +22,7 @@ class App extends PureComponent {
   }
 
   _renderApp() {
-    const {offerPlacesCount, offersArray} = this.props;
+    const {offerPlacesCount, offersArray, cityes} = this.props;
     const {activeId} = this.state;
 
     if (activeId < 0) {
@@ -31,6 +31,7 @@ class App extends PureComponent {
           offerPlacesCount={offerPlacesCount}
           offersArray={offersArray}
           onApartmentCardClick={this._handleApartmentCardClick}
+          cityes={cityes}
         />
       );
     }
@@ -67,7 +68,8 @@ App.propTypes = {
     price: PropTypes.number,
     srcImg: PropTypes.string
   })).isRequired,
-  onApartmentCardClick: PropTypes.func
+  onApartmentCardClick: PropTypes.func,
+  cityes: PropTypes.object.isRequired
 };
 
 export default App;
