+++ src/mocks/cityes.js
@@ -1,0 +1,10 @@
+const cityes = {
+  AMSTERDAM: [52.38333, 4.9],
+  PARIS: [48.856663, 2.351556],
+  COLOGNE: [50.930779, 6.938399],
+  BRUSSELS: [50.851309, 4.351718],
+  HAMBURG: [53.552645, 9.966287],
+  DUSSELDORF: [51.230569, 6.787428]
+};
+
+export default cityes;
